# docs/swagger.yaml
basePath: /
definitions:
...
  model.UserInfo:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      password:
        type: string
      sayHello:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  user.CreateRequest:
    properties:
      captchaId:
        type: string
      captchaValue:
        type: string
      email:
        type: string
      group:
        type: integer
      id:
        type: integer
      id_card:
        type: string
      is_super:
        type: boolean
      nickname:
        type: string
      oldPassword:
        type: string
      password:
        type: string
      picture:
        type: string
      state:
        type: integer
      username:
        type: string
      users:
        items:
          type: integer
        type: array
    type: object
  user.CreateResponse:
    properties:
      username:
        type: string
    type: object
  user.ListRequest:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      username:
        type: string
    type: object
  user.SwaggerListResponse:
    properties:
      totalCount:
        type: integer
      userList:
        items:
          $ref: '#/definitions/model.UserInfo'
        type: array
    type: object
host: localhost:8083
info:
  contact:
    email: yljckh@gmail.com
    name: yeelone
    url: https://www.tdongli.com
  description: hr-server
  title: hr-server API
  version: "1.0"
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: 列出所有用户列表
      parameters:
      - description: List users
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.ListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"totalCount":1,"userList":[{"id":0,"username":"admin","random":"user
            ''admin'' get random string ''EnqntiSig''","password":"$2a$10$veGcArz47VGj7l9xN7g2iuT9TF21jLI1YGXarGzvARNdnt4inC9PG","createdAt":"2018-05-28
            00:25:33","updatedAt":"2018-05-28 00:25:33"}]}}'
          schema:
            $ref: '#/definitions/user.SwaggerListResponse'
      summary: List the users in the database
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Add a new user
      parameters:
      - description: Create a new user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/user.CreateResponse'
      summary: Add new user to the database
      tags:
      - user
...
swagger: "2.0"